Change the way we lock the controls.
Unfortunately displaying an empty GuiScreen does not really work, because
the minecraft HUD will always be drawn if a GuiScreen is active.
And since everyone probably wants to hide the HUD while making timelapses,
we need to be able to hide it.
So now we lock the controls by calling setIngameNotInFocus() and
Keybinding.unPressAllKeys() every tick in the PlayerTickHandler.