Change a lot of the instance handling/storage stuff. To be cleaned-up better.
This version mostly works already.
BUGS/issues: The motion is jittery/non-smooth,
and the control locking doesn't work properly, like it did in 1.5 and 1.6.
It lets through the initial key press for a split second, and if the mouse
is clicked and dragged, the screen moves a little bit.

And then the overall program structure could use some serious cleaning
up...