Changes to the tick/event stuff to fix some problems and timing bugs.
Moved the mouse locking and motion back to ClientTickEvent.
It seems to work if it is executed in the END phase.
The PlayerTickEvent would get executed both on the client and the server
side, and also it seems to fire twice per tick, with 15ms and 35ms
intervals.
The keyboard key locking is now done in the KeyEvent handler.

This commit includes commented out debug code.